Persistence instrumentation and unit tests have been updated.
